Improve code placements
