+++ homeassistant/components/tcp/binary_sensor.py
@@ -8,8 +8,8 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType
 
+from .common import TCP_PLATFORM_SCHEMA, TcpSensor
 from .const import CONF_VALUE_ON
-from .sensor import PLATFORM_SCHEMA as TCP_PLATFORM_SCHEMA, TcpSensor
 
 PLATFORM_SCHEMA: Final = TCP_PLATFORM_SCHEMA
